apps/backend-fastapi/requirements.txt
uvicorn
fastapi
geoalchemy2
shapely
python-dotenv